requirements.txt
aiohttp
discord.py
pandas
ta
ta-lib
